refactor(remify): turn the remify function into a module
use it for converting pixel values into rem, not just for font size but also for layout parameters
